+++ websockets/py35/_test_client_server.py
@@ -5,7 +5,7 @@ import unittest
 
 from ..client import *
 from ..server import *
-from ..test_client_server import handler
+from ..test_client_server import get_server_uri, handler
 
 
 class ContextManagerTests(unittest.TestCase):
@@ -18,24 +18,24 @@ class ContextManagerTests(unittest.TestCase):
         self.loop.close()
 
     def test_client(self):
-        server = serve(handler, 'localhost', 8642)
-        self.server = self.loop.run_until_complete(server)
+        start_server = serve(handler, 'localhost', 0)
+        server = self.loop.run_until_complete(start_server)
 
         async def run_client():
-            async with connect('ws://localhost:8642/') as client:
+            async with connect(get_server_uri(server)) as client:
                 await client.send("Hello!")
                 reply = await client.recv()
                 self.assertEqual(reply, "Hello!")
 
         self.loop.run_until_complete(run_client())
 
-        self.server.close()
-        self.loop.run_until_complete(self.server.wait_closed())
+        server.close()
+        self.loop.run_until_complete(server.wait_closed())
 
     def test_server(self):
         async def run_server():
-            async with serve(handler, 'localhost', 8642):
-                client = await connect('ws://localhost:8642/')
+            async with serve(handler, 'localhost', 0) as server:
+                client = await connect(get_server_uri(server))
                 await client.send("Hello!")
                 reply = await client.recv()
                 self.assertEqual(reply, "Hello!")